A Dark Theme for PeaZip
Just put this file into C:\Program Files\PeaZip\res\share\themes and then select it from the program, and you're good to go. (I tried to make the text white to make it more visible in the dark theme, but it's not possible as of right now unless you know how to code in this programming language that I don't even know. Here's the repository if you're interested in doing that: https://github.com/peazip/PeaZip).